#56 Moved all previously shared core to more sensible locations and moved all OnceCells to LazyCells
