+++ rustyms/src/formula/mod.rs
@@ -2,8 +2,11 @@ mod pro_forma;
 mod psi_mod;
 #[macro_use]
 mod structure;
+mod formula;
 mod resid;
+mod search;
 mod unimod;
 mod xlmod;
 
+pub use search::*;
 pub use structure::*;
