+++ rustyms/src/formula/resid.rs
