+++ rustyms/src/glycan/mod.rs
@@ -1,13 +1,18 @@
 //! Handle glycan related issues, access provided if you want to work with glycans on your own.
 
+mod glycan;
 mod glycan_structure;
+mod lists;
 mod monosaccharide;
+mod position;
 mod positioned_structure;
 #[cfg(feature = "glycan-render")]
 mod render;
 
+pub use glycan::*;
 pub use glycan_structure::*;
-pub use monosaccharide::*;
+pub(crate) use lists::*;
+pub use position::GlycanPosition;
 pub use positioned_structure::*;
 #[cfg(feature = "glycan-render")]
 pub use render::{GlycanDirection, GlycanRoot, GlycanSelection, RenderedGlycan};
