+++ rustyms/src/glycan/monosaccharide.rs
@@ -1,15 +1,16 @@
 //! Handle monosaccharides
 
+use itertools::Itertools;
+
 use crate::{
     fragment::{DiagnosticPosition, Fragment, FragmentType},
     model::GlycanModel,
     molecular_charge::CachedCharge,
     system::usize::Charge,
-    AminoAcid, FragmentationModel, Multi,
+    AminoAcid, Chemical, FragmentationModel, MolecularFormula, Multi, SequencePosition,
 };
 
-include!("../shared/glycan.rs");
-include!("../shared/glycan_lists.rs");
+use super::glycan::MonoSaccharide;
 
 impl MonoSaccharide {
     /// Generate the composition used for searching on glycans
@@ -190,7 +191,18 @@ impl MonoSaccharide {
 #[cfg(test)]
 #[expect(clippy::missing_panics_doc)]
 mod tests {
-    use super::*;
+    use itertools::Itertools;
+
+    use crate::{
+        glycan::{
+            glycan::{
+                BaseSugar, Configuration, GlycanSubstituent, HexoseIsomer, MonoSaccharide,
+                PentoseIsomer,
+            },
+            lists::GLYCAN_PARSE_LIST,
+        },
+        Chemical,
+    };
 
     #[test]
     fn pro_forma_compliance() {
@@ -222,7 +234,7 @@ mod tests {
         ];
         for (name, formula) in cases {
             assert_eq!(
-                glycan_parse_list()
+                GLYCAN_PARSE_LIST
                     .iter()
                     .find(|p| p.0 == *name)
                     .unwrap_or_else(|| panic!("Assumed {name} would be defined"))
