+++ rustyms/src/glycan/positioned_structure.rs
@@ -4,10 +4,9 @@ use std::{hash::Hash, ops::RangeInclusive};
 use itertools::Itertools;
 use serde::{Deserialize, Serialize};
 
-use super::MonoSaccharide;
 use crate::{
     formula::{Chemical, MolecularFormula},
-    fragment::{Fragment, FragmentType, GlycanBreakPos, GlycanPosition},
+    fragment::{Fragment, FragmentType, GlycanBreakPos},
     model::GlycanModel,
     molecular_charge::CachedCharge,
     system::usize::Charge,
@@ -15,6 +14,8 @@ use crate::{
 };
 
 use crate::uom::num_traits::Zero;
+
+use super::{glycan::MonoSaccharide, position::GlycanPosition};
 
 /// The index in the branches as stored in the structure
 pub type GlycanBranchIndex = usize;
