+++ rustyms/src/identification/novob.rs
@@ -16,59 +16,54 @@ use super::{
     BoxedIdentifiedPeptideIter, SequenceElement,
 };
 
-use std::sync::OnceLock;
+use std::sync::LazyLock;
 
 static NUMBER_ERROR: (&str, &str) = (
     "Invalid NovoB line",
     "This column is not a number but it is required to be a number in this format",
 );
 
-static PARAMETERS_LOCK: OnceLock<SloppyParsingParameters> = OnceLock::new();
-
 /// Global parsing parameters
-#[expect(clippy::missing_panics_doc)] // These modifications exist
-fn parameters() -> &'static SloppyParsingParameters {
-    PARAMETERS_LOCK.get_or_init(|| SloppyParsingParameters {
-        custom_alphabet: vec![
-            (
-                b's',
-                SequenceElement::new(AminoAcid::Serine.into(), None)
-                    .with_simple_modification(Ontology::Unimod.find_id(21, None).unwrap()),
-            ),
-            (
-                b't',
-                SequenceElement::new(AminoAcid::Tyrosine.into(), None)
-                    .with_simple_modification(Ontology::Unimod.find_id(21, None).unwrap()),
-            ),
-            (
-                b'y',
-                SequenceElement::new(AminoAcid::Threonine.into(), None)
-                    .with_simple_modification(Ontology::Unimod.find_id(21, None).unwrap()),
-            ),
-            (
-                b'n',
-                SequenceElement::new(AminoAcid::Asparagine.into(), None)
-                    .with_simple_modification(Ontology::Unimod.find_id(7, None).unwrap()),
-            ),
-            (
-                b'q',
-                SequenceElement::new(AminoAcid::Glutamine.into(), None)
-                    .with_simple_modification(Ontology::Unimod.find_id(7, None).unwrap()),
-            ),
-            (
-                b'C',
-                SequenceElement::new(AminoAcid::Cysteine.into(), None)
-                    .with_simple_modification(Ontology::Unimod.find_id(6, None).unwrap()),
-            ),
-            (
-                b'm',
-                SequenceElement::new(AminoAcid::Methionine.into(), None)
-                    .with_simple_modification(Ontology::Unimod.find_id(35, None).unwrap()),
-            ),
-        ],
-        ..Default::default()
-    })
-}
+static PARAMETERS: LazyLock<SloppyParsingParameters> = LazyLock::new(|| SloppyParsingParameters {
+    custom_alphabet: vec![
+        (
+            b's',
+            SequenceElement::new(AminoAcid::Serine.into(), None)
+                .with_simple_modification(Ontology::Unimod.find_id(21, None).unwrap()),
+        ),
+        (
+            b't',
+            SequenceElement::new(AminoAcid::Tyrosine.into(), None)
+                .with_simple_modification(Ontology::Unimod.find_id(21, None).unwrap()),
+        ),
+        (
+            b'y',
+            SequenceElement::new(AminoAcid::Threonine.into(), None)
+                .with_simple_modification(Ontology::Unimod.find_id(21, None).unwrap()),
+        ),
+        (
+            b'n',
+            SequenceElement::new(AminoAcid::Asparagine.into(), None)
+                .with_simple_modification(Ontology::Unimod.find_id(7, None).unwrap()),
+        ),
+        (
+            b'q',
+            SequenceElement::new(AminoAcid::Glutamine.into(), None)
+                .with_simple_modification(Ontology::Unimod.find_id(7, None).unwrap()),
+        ),
+        (
+            b'C',
+            SequenceElement::new(AminoAcid::Cysteine.into(), None)
+                .with_simple_modification(Ontology::Unimod.find_id(6, None).unwrap()),
+        ),
+        (
+            b'm',
+            SequenceElement::new(AminoAcid::Methionine.into(), None)
+                .with_simple_modification(Ontology::Unimod.find_id(35, None).unwrap()),
+        ),
+    ],
+    ..Default::default()
+});
 
 format_family!(
     /// The format for any NovoB file
@@ -99,7 +94,7 @@ format_family!(
                 location.full_line(),
                 location.location.clone(),
                 custom_database,
-                parameters()
+                &PARAMETERS
             )).transpose();
 
         score_reverse: f64, |location: Location, _| location.parse::<f64>(NUMBER_ERROR);
@@ -109,7 +104,7 @@ format_family!(
                 location.full_line(),
                 location.location.clone(),
                 custom_database,
-                parameters(),
+                &PARAMETERS,
             )).transpose();
     }
     optional { }
