+++ rustyms/src/imgt/fancy.rs
@@ -1,4 +1,4 @@
-pub use super::shared::*;
+pub use super::*;
 use std::fmt::Display;
 use std::fmt::Write;
