--- rustyms/src/imgt/shared/mod.rs
@@ -1,4 +1,0 @@
-mod regions;
-mod species;
-pub use regions::*;
-pub use species::*;
