+++ rustyms/src/lib.rs
@@ -35,7 +35,6 @@ pub mod aminoacid;
 mod checked_aminoacid;
 mod element;
 pub mod error;
-mod formula_search;
 pub mod fragment;
 pub mod glycan;
 mod isobaric_sets;
@@ -46,7 +45,6 @@ mod mass_mode;
 pub mod model;
 pub mod modification;
 mod molecular_charge;
-#[path = "shared/multi.rs"]
 mod multi;
 mod mzpaf;
 mod neutral_loss;
@@ -59,7 +57,6 @@ mod protease;
 mod rand;
 pub mod rawfile;
 mod sequence_element;
-#[path = "shared/sequence_position.rs"]
 mod sequence_position;
 pub mod spectrum;
 pub mod system;
@@ -69,7 +66,6 @@ pub use aminoacid::{AminoAcid, IsAminoAcid};
 pub use checked_aminoacid::CheckedAminoAcid;
 pub use element::*;
 pub use formula::*;
-pub use formula_search::find_formulas;
 pub use fragment::Fragment;
 pub use isobaric_sets::{building_blocks, find_isobaric_sets};
 pub use mass_mode::MassMode;
